Abilities and Wave Spawn
Added the roar and stomp abilities. Roar works fine but stomp is having Concurrent troubles when removing enemies. Line 1381
Enemies now stop for about 2 seconds when the roar ability is pressed, with the exception of the newly added barbarians who move at double speed for the duration.
Wave Spawning more or less works but for some reason one of the footmen spawn before the thieves, someone look into that. Line 931
Added a method addEnemy to streamline having to add enemies in the game. Line 1597