Wonky Fireball
Fireball ability is now able to be primed, it waits for the player to press a button then creates a new fireball projectile.
It should home in on the x,y of the input but instead flies off towards the right of the screen in a curve. The curve is always the same when fireballs are cast to the same point. (Funny but not what we want.)
Also Fireball ability doesn't toggle off if the fireball ability button is pressed again.